Conflicts:
	scilab/modules/helptools/macros/xmltoformat.sci
	scilab/modules/output_stream/sci_gateway/cpp/sci_diary.cpp
	scilab/modules/scicos/src/c/scicos.c
	scilab/modules/spreadsheet/includes/gw_spreadsheet.h
	scilab/modules/spreadsheet/sci_gateway/c/gw_spreadsheet.c
	scilab/modules/spreadsheet/sci_gateway/spreadsheet_gateway.xml
	scilab/modules/spreadsheet/src/c/csvRead.c
	scilab/modules/spreadsheet/src/c/csvWrite.c
	scilab/modules/spreadsheet/src/c/csv_strsubst.c
	scilab/modules/string/Makefile.am
	scilab/modules/string/Makefile.in
	scilab/modules/string/includes/gw_string.h
	scilab/modules/string/sci_gateway/c/gw_string.c
	scilab/modules/string/sci_gateway/c/sci_isnum.c
	scilab/modules/string/src/c/string.vcxproj
	scilab/modules/ui_data/src/cpp/UpdateBrowseVar.cpp

+++ scilab/modules/boolean/sci_gateway/cpp/sci_find.cpp
@@ -45,7 +45,7 @@ types::Function::ReturnValue sci_find(types::typed_list &in, int _iRetCount, typ
             return types::Function::Error;
         }
 
-        iMax = in[1]->getAs<types::Double>()->get(0);
+        iMax = (int)in[1]->getAs<types::Double>()->get(0);
     }
 
     int* piIndex = 0;
@@ -84,7 +84,7 @@ types::Function::ReturnValue sci_find(types::typed_list &in, int _iRetCount, typ
     else if(in[0]->isSparse())
     {
         types::Sparse* pSP = in[0]->getAs<types::Sparse>();
-        int iNNZ = pSP->nonZeros();
+        int iNNZ = (int)pSP->nonZeros();
         int iRows = pSP->getRows();
         double* pRows = new double[iNNZ * 2];
         
@@ -93,14 +93,14 @@ types::Function::ReturnValue sci_find(types::typed_list &in, int _iRetCount, typ
 
         for(int i = 0 ; (iMax == -1 || iValues < iMax) && i < iNNZ ; i++)
         {
-            piIndex[iValues] = (pCols[i]-1) * iRows + (pRows[i]-1);
+            piIndex[iValues] = ((int)pCols[i]-1) * iRows + ((int)pRows[i]-1);
             iValues++;
         }
     }
     else if(in[0]->isSparseBool())
     {
         types::SparseBool* pSB = in[0]->getAs<types::SparseBool>();
-        int iNNZ = pSB->nbTrue();
+        int iNNZ = (int)pSB->nbTrue();
         int iRows = pSB->getRows();
         double* pRows = new double[iNNZ * 2];
         
@@ -109,7 +109,7 @@ types::Function::ReturnValue sci_find(types::typed_list &in, int _iRetCount, typ
 
         for(int i = 0 ; (iMax == -1 || iValues < iMax) && i < iNNZ ; i++)
         {
-            piIndex[iValues] = (pCols[i]-1) * iRows + (pRows[i]-1);
+            piIndex[iValues] = ((int)pCols[i]-1) * iRows + ((int)pRows[i]-1);
             iValues++;
         }
     }
